sound using the pc speaker

+++ kernel/memory/memory.h
@@ -1,4 +1,4 @@
 #include <stdint.h>
 void memset(void* start,uint_64 value,uint_64 num);
 void memcpy(void* dest,void* src,uint_64 num);
-int memcmp(const void* aptr,const void* bptr,size_t n);
+int memcmp(const void* aptr,const void* bptr,unsigned int n);
